Further modulation, and electric attack.
