+++ index.html
@@ -23,6 +23,8 @@
 
 
     <!-- <script src="../addons/p5.sound.js"></script> -->
+    <script src="player.js"></script>
+    <script src="attacks.js"></script>
     <script src="sketch.js"></script>
 </head>
